Refined type on currentTableSortDescription and moved checkmobile hook
